define add subject and todo using DB

+++ app/src/main/java/com/kania/todotree/data/ISubjectData.java
@@ -1,7 +1,7 @@
 package com.kania.todotree.data;
 
 public interface ISubjectData {
-    int NON_ID = -1;
+    long NON_ID = -1;
 
     long getId();
 }
